Update 1342-number-of-steps-to-reduce-a-number-to-zero.java
Advanced Concepts:

Binary representation of integers 'xxxxxxxx' (e.g 15 in binary format is '00001111'
Bitwise Shift Operators  '>>' ('>> 1' mean bits shift one position to the right, which equals the number divided by 2)
Bitwise Logical Operators  '&'
Bitmasks (It is '1' in this case. Using it with '&' to decide if the number is even or odd. If the result is 0 then the number is even, 1 then odd)